+++ src/State/AbstractKeywordState.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace umulmrum\JsonParser\State;
+namespace Umulmrum\JsonParser\State;
 
-use umulmrum\JsonParser\DataSource\DataSourceInterface;
-use umulmrum\JsonParser\InvalidJsonException;
+use Umulmrum\JsonParser\DataSource\DataSourceInterface;
+use Umulmrum\JsonParser\InvalidJsonException;
 
 /**
  * @internal