+++ src/State/DocumentEndState.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace umulmrum\JsonParser\State;
+namespace Umulmrum\JsonParser\State;
 
-use umulmrum\JsonParser\DataSource\DataSourceInterface;
+use Umulmrum\JsonParser\DataSource\DataSourceInterface;
 
 /**
  * @internal