+++ src/State/States.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace umulmrum\JsonParser\State;
+namespace Umulmrum\JsonParser\State;
 
 /**
  * Holds instances of all defined JSON states, so that only a single instance of each state is required, avoiding